fix(aio/upgrade-cli-src): print errors to the console
Previously, errors would cause the program to exit, but the error
messages were not shown in the log output. They were either swallowed or
reported as issues if running on CI.

This commit ensures that the error messages are surfaced, by always
printing them to the console (in addition to any other actions).