fix(aio/upgrade-cli-src): avoid race-condition with other triaging bots on PR milestone
Previously, due to a race-condition with other triaging bots (such as
`ngbot`), it was possible for the milestone to be overwritten to a
generic value (e.g. `Backlog`).

This commit tries to avoid such race-conditions by waiting for 10
seconds before setting the appropriate milestone. Normally, any
automatic triaging should have happened in that period.