fix(aio/upgrade-cli-src): correctly match the stable branch for `--branch=stable`
Previously, when `--branch` was set to `stable`, the script would look
for the highest patch branch in the repo. When a version is in RC, the
highest patch branch does not correspond to the stable version (i.e. the
one marked as `latest` on npm).

This commit fixes the script to correctly compute the stable branch as
the highest patch branch whose major version matches the major version
marked as `latest` on npm.