feat(aio/upgrade-cli-src): support upgrading the patch branch
Previously, only updating `master` and the stable branch were supported.
This commit adds support for checkign and upgrading the patch branch,
i.e. the highest branch that corresponds to a released version (e.g.
`3.2.x`).

_Background:_
Usually, the highest patch branch is the stable branch, but not while a
version is in RC (in which case the stable branch is the branch
corresponding to an older patch branch). While the patch branch is not
deployed to angular.io while in RC, it still makes sense to keep it
up-to-date, because it will become the new stable branch at some point.