feat(aio/upgrade-cli-src): support upgrading for a specific Angular major version
Most of the time, upgrading master, stable and the highest patch branch
(in case it is not the same as the stable one - i.e. during RC periods)
is enough. However, it might occasionally be necessary to upgrade an
older branch.

This commit adds support for targeting the branch that corresponds to a
specific major Angular version, such as version 3 (pun intended :P):
`--branch=3`