fix(aio/upgrade-cli-src): use name of target branch in created branch name
Previously, the branch spec was used instead (e.g. `stable` or `3`).
This did not provide enough information for future executions to
determine whether a new PR is needed, since the branch corresponding to
a branch spec could have changed since the original PR.

This commit eliminates the ambiguity by using the actual name of the
target branch in the created branch name.