fix(aio/upgrade-cli-src): fetch full branch when creating local branch
Previouly, we only fetched the last commit of a branch when creating a
local branch in order to save time. However, before creating a pull
request, we had to fetch the full history with `fetch --unshallow`.

Since there is currently no scenario in which we create a local branch
without creating a pull request, this commit removes the need for
unshallowing by fetching the branch from the beginning.

I am also hoping that this will fix errors such as #56, which is
possibly caused by the fact that unshallowed working copies behave
subtly differently than full working copies.